refactor(PostCard): レビュー対応 - 型・util分離とisLoading props削除
- PostCardProps を types.ts に切り出し責務を分離
- formatRelativeTime を src/utils/ に移動しどこからでも再利用可能に
- isLoading props を削除（Suspenseに委ねる設計に変更）
- PostCardSkeleton は Suspense fallback 用として維持
