+++ src/utils/formatRelativeTime.ts
@@ -1,0 +1,30 @@
+/**
+ * 日付を受け取り、「たった今」や「◯分前」といった相対時間文字列に変換するユーティリティ関数
+ */
+export const formatRelativeTime = (date: string | Date | undefined): string => {
+  if (!date) return "";
+
+  const targetDate = new Date(date);
+  const now = new Date();
+
+  const diffInMs = now.getTime() - targetDate.getTime();
+  const diffInMinutes = Math.floor(diffInMs / (1000 * 60));
+  const diffInHours = Math.floor(diffInMs / (1000 * 60 * 60));
+  const diffInDays = Math.floor(diffInMs / (1000 * 60 * 60 * 24));
+  const diffInMonths = Math.floor(diffInDays / 30);
+  const diffInYears = Math.floor(diffInDays / 365);
+
+  if (diffInMinutes < 1) {
+    return "たった今";
+  } else if (diffInMinutes < 60) {
+    return `${diffInMinutes}分前`;
+  } else if (diffInHours < 24) {
+    return `${diffInHours}時間前`;
+  } else if (diffInDays < 30) {
+    return `${diffInDays}日前`;
+  } else if (diffInMonths < 12) {
+    return `${diffInMonths}ヶ月前`;
+  } else {
+    return `${diffInYears}年前`;
+  }
+};
